feat(ELE-2438): consolidate spatial providers under geo/providers/ (D3)
Structural moves only \u2014 interface unification deferred to a later epic.

## Moves
- geo/boundary_providers.py         -> geo/providers/boundary_providers.py
- geo/census_geocoder.py            -> geo/providers/census_geocoder.py
- geo/nces_download.py              -> geo/providers/nces_download.py
- data/redistricting_data_hub.py    -> geo/providers/redistricting_data_hub.py

## Back-compat
Deprecation shims at each old path re-export from the new location and
emit DeprecationWarning; remove in the next minor release.

geo/__init__.py lazy-import table updated to point directly at the new
provider locations (no internal shim triggering). data/__init__.py's
top-level RDH re-exports now pull from geo/providers/ so the data.*
facade stays quiet too.

## Not migrated in this PR
- geo/spatial_data.py (1,500+ lines mixing CensusDataSource, Government,
  and OSM sources) \u2014 needs a file split, deferred to interface-unification
  follow-up.
- geo/census/ subpackage (Census API client) \u2014 moves once the
  variable-registry refactor stabilizes.

## Verified
- Library imports cleanly
- 4 deprecation shims fire on old-path import
- 151 provider tests pass (boundary_providers, census_geocoder,
  nces_download, redistricting_data_hub)
