+++ tests/test_nces_service.py
@@ -65,7 +65,7 @@ class TestNCESPopulationService:
         mod = importlib.import_module("siege_utilities.geo.django.services.nces_service")
         NCESPopulationService = mod.NCESPopulationService
 
-        from siege_utilities.geo.nces_download import NCESDownloader
+        from siege_utilities.geo.providers.nces_download import NCESDownloader
 
         service = NCESPopulationService(cache_dir="/tmp/test")
         downloader = service._get_downloader()
